+++ src/js/features/constants.js
@@ -1,4 +1,5 @@
 export const sections = {
+  GLOBAL: 'global',
   DECK_VIEW: 'deck-view',
   DECK_BUILDER: 'deck-builder'
 }