+++ src/js/features/global-features/index.js
@@ -1,0 +1,5 @@
+import FutureFeatureOptIn from './future-feature-opt-in'
+
+export default [
+  FutureFeatureOptIn
+]