+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Changelog
 --------
 
+# unreleased
+* Add opt-in setting for future features
+
 # 0.1.0 - Pre-release
 * Add EDHRec Suggestions Button